requirements.txt
torch==2.2.2+cu118
torchvision==0.17.2+cu118
torchtext==0.17.2
torchdata==0.7.1
gradio
pillow
numpy<2.0.0
